// admcentro-web-main/src/interfaces/Icontracts.ts
// Generated by https://quicktype.io

import { IAssurance } from './IAssurance'
import { Idebt } from './IDebtsResponse'
import { IPerson } from './Iowners'
import { Iproperty } from './Iproperties'

export interface IcontractsResponse {
	results: number
	code: number
	status: string
	ok: boolean
	data: Contract[]
}

export interface Contract {
	id: number
	uuid: string
	PropertyId: number
	ClientId: number
	startDate: string
	endDate: string
	state: string
	description: string
	deposit: number
	booking: number
	createdAt: string
	updatedAt: string
	Client: IPerson
	Property: Iproperty
	PriceHistorials: IHistorialPrice[]
	DebtClients: Idebt[]
	DebtOwners: Idebt[]
	Assurances: IAssurance[]
	amount: number
	admFeesPorc: number
	currency: 'ARS' | 'USD'
	paymentType: "Fijo" | "Porcentual",
	adjustmentMonth: number,
}


export interface IHistorialPrice {
	id: number
	ContractId: number
	amount: number
	year: number
	percent: number
	createdAt: string
	updatedAt: string
	deletedAt: null | string
}